+++ packages/web/src/app/[lng]/(main)/community/[articleId]/components/ReplyItem.tsx
@@ -15,7 +15,7 @@ import { Icon } from '@/components/Icon';
 import { Flex } from '@/components/Layout';
 import { Loading } from '@/components/Loading';
 import { Spacing } from '@/components/Spacing';
-import { useModal } from '@/hooks/useModal';
+import useModal from '@/hooks/useModal/useModal';
 import { useBlockStore } from '@/store/useBlockStore';
 import cn from '@/utils/cn';
 import { getLocalCookie } from '@/utils/cookieController';
