+++ packages/web/src/app/[lng]/(main)/community/write/components/WriteHeader.tsx
@@ -7,7 +7,7 @@ import { IconButton } from '@/components/Button';
 import { Header } from '@/components/Header';
 import { Icon } from '@/components/Icon';
 import useAppRouter from '@/hooks/useAppRouter';
-import { useModal } from '@/hooks/useModal';
+import useModal from '@/hooks/useModal/useModal';
 
 export default function WriteHeader() {
   const { t } = useTranslation('community');
