+++ packages/web/src/app/[lng]/(main)/grouping/hooks/useMoreSheet.tsx
@@ -2,11 +2,12 @@
 import BlockDoneModal from '../components/BlockDoneModal';
 import ReportDoneModal from '../components/ReportDoneModal';
 import WarningModal from '../components/WarningModal';
+
 import { useDeleteArticle, useDeleteComment } from '@/apis/groups';
 import { useTranslation } from '@/app/i18n/client';
 import MoreBottomSheet from '@/components/Modal/MoreBottomSheet';
 import useAppRouter from '@/hooks/useAppRouter';
-import { useModal } from '@/hooks/useModal';
+import useModal from '@/hooks/useModal/useModal';
 import { useBlockStore } from '@/store/useBlockStore';
 
 type CommentId<T> = T extends 'comment' ? { commentId: number } : { commentId?: never };
