+++ packages/web/src/app/[lng]/(main)/meeting/participate/components/FeedbackContent.tsx
@@ -9,7 +9,7 @@ import { useTranslation } from '@/app/i18n/client';
 import { Button } from '@/components/Button';
 import { GroupingCard } from '@/components/Card';
 import { Spacing } from '@/components/Spacing';
-import { useModal } from '@/hooks/useModal';
+import useModal from '@/hooks/useModal/useModal';
 
 export default function FeedbackContent() {
   const { t } = useTranslation('meeting');
