+++ packages/web/src/app/[lng]/(main)/meeting/participate/components/WaitingContent.tsx
@@ -8,7 +8,7 @@ import { useTranslation } from '@/app/i18n/client';
 import { GroupingCard } from '@/components/Card';
 import { Divider } from '@/components/Divider';
 import { Spacing } from '@/components/Spacing';
-import { useModal } from '@/hooks/useModal';
+import useModal from '@/hooks/useModal/useModal';
 
 export default function WaitingContent() {
   const { t } = useTranslation('meeting');
