+++ packages/web/src/app/[lng]/(main)/meeting/participate/feedback/[groupId]/funnels/step1/Step1.tsx
@@ -13,7 +13,7 @@ import { Icon } from '@/components/Icon';
 import { Flex } from '@/components/Layout';
 import { Spacing } from '@/components/Spacing';
 import { Tag } from '@/components/Tag';
-import { useModal } from '@/hooks/useModal';
+import useModal from '@/hooks/useModal/useModal';
 
 interface Step1Props {
   onNextClick: () => void;
