+++ packages/web/src/app/[lng]/(main)/profile/mates/components/MatesDetail.tsx
@@ -10,7 +10,7 @@ import { Avatar } from '@/components/Avatar';
 import { Icon } from '@/components/Icon';
 import { Flex } from '@/components/Layout';
 import { Spacing } from '@/components/Spacing';
-import { useModal } from '@/hooks/useModal';
+import useModal from '@/hooks/useModal/useModal';
 
 export default function MatesDetail() {
   const { t } = useTranslation('meeting');
