+++ packages/web/src/app/[lng]/(main)/profile/verify/components/EmailForm.tsx
@@ -11,7 +11,7 @@ import { useTranslation } from '@/app/i18n/client';
 import { Button, ButtonGroup } from '@/components/Button';
 import { TextFieldController } from '@/components/TextField';
 import { regexr } from '@/constants/regexr';
-import { useModal } from '@/hooks/useModal';
+import useModal from '@/hooks/useModal/useModal';
 
 const defaultValues = {
   email: '',
