+++ packages/web/src/app/[lng]/(sub)/join/funnels/step2/Step2Component.tsx
@@ -8,7 +8,7 @@ import JoinTitleTextMessage from '../../components/JoinTitleTextMessage';
 import { useTranslation } from '@/app/i18n/client';
 import { LocalSuspenseErrorBoundary } from '@/components/ErrorBoundary';
 import { useDidMount } from '@/hooks/common/useDidMount';
-import { useModal } from '@/hooks/useModal';
+import useModal from '@/hooks/useModal/useModal';
 
 export default function Step2Component() {
   const { t } = useTranslation('join');
