+++ packages/web/src/app/[lng]/(sub)/join/funnels/step3/components/EmailForm.tsx
@@ -13,7 +13,7 @@ import { useTranslation } from '@/app/i18n/client';
 import { Button, ButtonGroup } from '@/components/Button';
 import { TextFieldController } from '@/components/TextField';
 import { regexr } from '@/constants/regexr';
-import { useModal } from '@/hooks/useModal';
+import useModal from '@/hooks/useModal/useModal';
 
 export default memo(function EmailForm() {
   const { t } = useTranslation('join');
