+++ packages/web/src/app/[lng]/(sub)/join/page.tsx
@@ -4,7 +4,7 @@ import JoinContextProvider from './components/JoinContext';
 import JoinFunnel from './funnels/JoinFunnel';
 
 import { Loading } from '@/components/Loading';
-import { ModalProvider } from '@/hooks/useModal';
+import ModalProvider from '@/hooks/useModal/ModalProvider';
 
 export default async function JoinPage() {
   return (
