+++ packages/web/src/components/Provider/QueryProvider.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable react/jsx-no-useless-fragment */
 'use client';
 
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
@@ -7,7 +6,7 @@ import { useState } from 'react';
 
 import type { StrictPropsWithChildren } from '@/types';
 
-import { useToast } from '@/hooks/useModal';
+import useToast from '@/hooks/useModal/useToast';
 
 export default function QueryProvider({ children }: StrictPropsWithChildren) {
   const { openToast } = useToast();
