--- packages/web/src/hooks/useModal/index.ts
@@ -1,3 +1,0 @@
-export { default as ModalProvider } from './ModalProvider';
-export { default as useModal } from './useModal';
-export { default as useToast } from './useToast';
