+++ packages/web/src/hooks/useModal/useToast.tsx
@@ -1,4 +1,5 @@
 import useModal from './useModal';
+
 import { Toast } from '@/components/Modal';
 
 export default function useToast() {
